+++ ASCOfficeOdfFile/src/odf/svg_parser.cpp
@@ -817,11 +817,11 @@ namespace svg_path
 		
 		boost::algorithm::split(points_str,rPolygonStatement, boost::algorithm::is_any_of(L" "), boost::algorithm::token_compress_on);
 
-		int last_x=0,last_y=0;
-		for (long i=0; i< points_str.size();i++)
+		int last_x = 0, last_y = 0;
+		for (size_t i = 0; i < points_str.size(); i++)
 		{
  			std::vector<std::wstring> point_str;
-			boost::algorithm::split(point_str,points_str[i], boost::algorithm::is_any_of(L","), boost::algorithm::token_compress_on);
+			boost::algorithm::split(point_str, points_str[i], boost::algorithm::is_any_of(L","), boost::algorithm::token_compress_on);
 
 			if (point_str.size()>0)last_x = boost::lexical_cast<int>(point_str[0]);
 			if (point_str.size()>1)last_y = boost::lexical_cast<int>(point_str[1]);
